Added seperate date-picker documentation component
