+++ component-library/src/app/app-routing.module.ts
@@ -24,13 +24,9 @@ import { NaseerComponent } from './pages/QA/naseer/naseer.component';
 import { AccessibilityDemoComponent } from './pages/QA/accessibility-demo/accessibility-demo.component';
 import { AccessibilityDemoPreviousPageComponent } from './pages/QA/accessibility-demo/accessibility-demo-previous-page/accessibility-demo-previous-page.component';
 import { AccessibilityDemoNextPageComponent } from './pages/QA/accessibility-demo/accessibility-demo-next-page/accessibility-demo-next-page.component';
-import { ContactComponent } from './pages/contact/contact.component';
-import { BannerDocCodeComponent } from './pages/banner-documentation/banner-doc-code.component';
 import { ButtonDocCodeComponent } from './pages/button-documentation/button-doc-code.component';
-import { AccessibilityDemoModule } from './pages/QA/accessibility-demo/accessibility-demo.module';
 import { BobbyComponent } from './pages/QA/bobby/bobby.component';
 import { InputDocumentationComponent } from './pages/input-documentation/input-documentation.component';
-import { InputDocCodeComponent } from '@app/pages/input-documentation/input-doc-code.component';
 import { RequestFormComponent } from './pages/request-form/request-form.component';
 import { BannerDocumentationComponent } from './pages/banner-documentation/banner-documentation.component';
 import { DatePickerDocumentationComponent } from './pages/date-picker-documentation/date-picker-documentation.component';
