+++ component-library/src/app/pages/date-picker-documentation/date-picker-doc-code.component.scss
@@ -1,0 +1,5 @@
+.datePicker-container {
+  max-width: 320px;
+  padding-top: 54px;
+  padding-bottom: 54px;
+}
