--- component-library/src/app/pages/date-picker-documentation/date-picker-documentation.component.scss
@@ -1,5 +1,0 @@
-.datePicker-container {
-  max-width: 320px;
-  padding-top: 54px;
-  padding-bottom: 54px;
-}
